fix case where a redirect with params yields garbage in url
A case can arise where a GET request, with params,  receives a 302
redirect and the subsequent request has param garbase appended to the
url.

Example:

    client> GET /some/url?id=1 HTTP/1.1
    client> Host: example.com

    server< HTTP/1.1 302 Moved Temporarily
    server< Location: http://example.com/some/url?id=1&extra_data=2

    ## this is the bad request right here. Note the [('id', 1)] stuck on
    # the end
    client> GET /some/url?id=1&extra_data=2&[('id', 1)] HTTP/1.1.
    client> Host: example.com

    # the server of course doesn't like it
    server< HTTP/1.0 400 Bad request.

This issue is caused by the conditional branch in _encode_params for the
second request not urlencoding the list of tuples for the
Request._enc_params attribute.

github note: possible fix for kennethreitz/requests#157