Added testcase showing a non-standard unicode character (in this case, A-umlaut), returned by the server in a redirect, breaks py27 requests.
I have a class of urls which I need to get in py27 which break,
because the server returns a _relative redirect_ with a high unicode
character in it.  Because of how str and unicode are re-bound in py_27
mode for requests, this means that unicode is being called on an
object that does not implicitly convert to unicode (it has an ord >
127).

The other fix is a small hack to remove the unicode-literal u from the
file so it can be run on python3.