Keeping track of setup environments for python 2 and 3 separate was driving me nuts, so I've added support for testing with tox and nosetests.
If you have tox installed, it should build requests into a virtualenv
for each environment (currently py27 and py32) and then run the tests.

Currently, this exposes that gevent is outside of the requests setup
(by design) so it needs to be included seperately for
test_requests_async.py to work.  Gevent doesn't exist for Python 3.

And envoy is not in the requirements of setup, despite being
referenced in the test cases.