Make sure form data are populated after a redirect
Copy the form data into the new request after a redirect.  The new request
can send the data via a POST or PUT after the redirect is followed.

A unit test for this could not be created as the response is assigned
the original request before the build() function returns.

requests/requests/models.py
+ Copied the data from the redirected request into the new request
    in the build() function.