Re-added the response attribute (now assigned with hook in Session's send method).
Added tests to verify that there is no response attribute in a
Session before a request is made, that the response object returned
by Session's request call is equal to the Session's request attribute,
and that the response attribute changes compared to the original
returned response when a new Session request is made to a different
resource.

Added my name to the AUTHORS list.