Avoid two requests on a TypeError
Rather than test behavior via try: except, inspect python version
number.  This may not be a huge issue in real life.  I noticed it
using wsgi-intercept where an intercepted app raised a
TypeError (which got bubbled up the stack to this catch).

If the purpose of this code is as the comments say, checking
sys.version_info is probably a more efficient and predictable.  But
maybe more is happening here than I see.

No test included, this is an informational PR! Cheers!