Suppress the warning if verify=False
Setting verify to False differs from verify=None in being explicitly set,
and therefore the warning being undesirable.

See this for the a proposed alternative that involves importing
implementation details of Requests onto the application:
 https://github.com/kennethreitz/requests/issues/2214#issuecomment-72941896

This fix avoids deep imports and pulling the knowledge of internals of
Requests into the application.

Fixes #2214