Support non-ASCII status reason as per RFC2616
RFC2616 specifies that header values and reason phrases are encoded in
in ISO-8859-1 or RFC2047 (MIME for text). RFC7230, which obsoletes
RC2616, recommends treating non-ASCII header values as 'opaque data'. In
the case of the reason phrase, we should try to decode it due to its
possible localization and intent for display to the user.

Fixes kennethreitz/requests#3009