Allow sending an unprepared request via session.
Seperate the parts of session that create a request from parameters with
the part that sends a request. From a user perspective I want to use and
pass around a request object rather than lots of parameters. However if
I do that I have to manually prepare it, merge the envrionment and
proxy settings and then send it.

Given that a user can make a request object it should be easier to use
one.