Enable sessions to bypass creating/parsing cookies.
We use requests for some service-to-service communication, neither side
of which utilizes cookies.  When we profiled our services, we see ~13%
of the time spend sending requests is either creating cookie jars or
extracting cookies to jars.  We explicitly don't care about cookies.

This PR will enable us to disable cookies for a given `requests.Session`
by setting `session.cookies = None`.  If `sessions.cookies` is None, we
won't create any cookies for the request nor will we parse them on the
response.