When Location decoding fails, fall back to original
Issue #3888 correctly identified Location headers as *usually* containing UTF-8
codepoints (when not correctly URL encoded), but this is not always the case.
For example the URL
http://www.finanzen.net/suchergebnis.asp?strSuchString=DE0005933931 redirects
to `b'/etf/ishares_core_dax\xae_ucits_etf_de'`, containing the Latin-1 byte for
the ® character.

If UTF-8 decoding fails, it is better to fall back to the original.