Handle double-quoted redirects from misconfigured servers
Some misconfigured servers will take an already-quoted URI and
double-quote parts of it when returning a redirect response.  What
this does is to transform '%' characters in quoted entities into
'%25'.  When the client then attempts to make use of this location, it
predictably does not work as intended.  We attempt to detect this
misconfiguration by counting substrings in the original URI and in the
redirect URI.  If the original URI has zero '%25' substrings, we count
the number of '%' signs in it.  Then we count the number of '%25'
substrings in the redirect URI.  If the redirect URI contains as many
'%25' substrings as the request URI contained '%' signs, we unquote
the redirect URI and pass it along for further processing

Sponsored by: Micron Technology, Inc.