Add read() to Response object.
The traditional 'read' interface is useful in some cases.
At very least it may be easier for code that expectes file-handle
like objects to just add requests support.

Due to gzip/deflate content-encoding it is not really possible for
the caller to use just use Response.raw.read().  If gzip was used
that would return compressed bytes.