Including a `__main__.py` file
Python supports the inclusion of a `__main__.py` file withing ones packages. I though it might be nice to include such a  file within requests to handle the more common tasks from the command line e.g. `GET`, `HEAD`, `PUT` and `POST`. I have included an example for both `HEAD` and `GET` though I'm not certain of their utility. 

More often then not it is necessary to download a file from a given url and I have added a `download` option to assist with that as well.