Update docs with for passing an array as a param
When passing an array for a `GET` request, some servers expect the key to be suffixed with a `[]`
to parse the URL correctly. Adding a note in the documentation for this case.

Further, Is it a good idea to make this behavior a default one? Concatenate a `[]` to the end of all keys
whose value is an array. Would love to hear your thoughts!