Fix output of test_lowlevel tests in case of timeout
The tests in test_lowlevel.py use the testserver.server.Server thread
object.

When using the object as a context manager, we wait until the server
thread is ready with a timeout.

Unfortunately, if the timeout is reached, we would not propagate the
error. Instead, we would return from the __enter__ method and access the
"with" body with an uninitialized value for port (i.e., port = 0).

Therefore, the tests fail with:

E           urllib3.exceptions.NewConnectionError:
<urllib3.connection.HTTPConnection object at 0x7f068405aec0>: Failed to
establish a new connection: [Errno 111] Connection refused

Reading this error does not make it obvious that this is the result of a
timeout.

Fixed by raising an error in case of timeout.

Signed-off-by: Olivier Gayot <olivier.gayot@canonical.com>