Document that Response.iter_lines is broken.
This method is broken in a couple of significant ways:

- If delimiter is None, and if one chunk happens to end with \r and
  the following chunk begins with \n, this method will yield a bogus
  blank line at that point.

- If delimiter is a string, and if a chunk happens to end with that
  string, this method will yield a bogus blank line at that point.

This is broken both with and without decode_unicode=True, and it is
broken both with and without stream=True.  (Setting stream=True makes
the problem worse, because then the chunk boundaries are partially
dependent on server or network I/O chunking.)

Unless and until this behavior is fixed, this method should not be
used.