Minor quirk: more straightforward __iter__ for CaseInsensitiveDict
With the actual CaseInsensitiveDict.__iter__ method, both the key and value are extracted but only the key is used. With this minor change we directly loop over the internal dict's keys (which is the default for the dict's __iter__). The difference is minimal, just 4 less bytecode instructions: https://godbolt.org/z/saT7Me3vb (uncomment either __iter__ impl.)