Don't normalise or double-escape urls
`requote_uri` was added in 2013 in #1361 to resolve #1360.

But in 2019 this was also solved in `urllib3` in
https://github.com/urllib3/urllib3/pull/1647.

Not only does the `urllib3` implementation more closely match what
all other clients are doing.  It also means that we are double-encoding
invalid uris.  If the redirect is invalid we're in unchartered territory
but the path that `requests` uses has no chance to decode to the same
that any other client uses, even to the most lenient server.

Simply drop `requote_uri`, as `urllib3`'s `parse_url` will always handle
invalid urls.