Ignore Local Version in Compatibility Check
According to PEP 440 guidelines, the local version identifier may use the period
character, however, doing so causes the compatibility check to fail as the
version parsing splits on periods and unpacks to the results assuming the string
follows the major, minor, patch convention. Ignoring the local version string
increases the likelihood that the assumption holds.

Closes: #7035

Signed-off-by: Bryan Fraschetti <bryan.fraschetti@canonical.com>